==> 01_Analyzing_CSV.ipynb <==
## (01) Analyse des deux documents format CSV fournis par l'agence Ecommerce

##### *Import des trois libraries nécessaires pour l'analyse d'un document CSV*

import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

---

### Lecture des fichiers au format CSV (1)

---

##### *Document numéro 1 : Datafinity Electronics Product Data*

pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

df1 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

### Nombre de lignes et colonnes

df1.shape

---

### Affichage du nom des colonnes

df1.columns

---

### Affichage des infos principales du dataframe

df1.info()

---

### Count des valeurs les plus présentes dans "Manufacturer"

df1["manufacturer"].value_counts()

---

### Lecture des fichiers au format CSV (2)

---

##### *Document numéro 2 : Datafinity Electronics Product Data Pricing*

pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

df2 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

### Nombre de lignes et colonnes

df2.shape

---

### Affichage du nom des colonnes

df2.columns

---

### Affichage des infos principales du dataframe

df2.info()

---
