==> 03_Create_CSV_tables.ipynb <==
## (03) Création des différents CSV nécessaires à l'importation des données

##### *Import des libraries nécessaires pour le nettoyage et l'exportation des différents CSV nécessaires à la création de la BDD*

import pandas as pd

---

## 1 - Création du CSV pour la table "Product"

#### *Document numéro 1 : Datafinity Electronics Product Data*

df1 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

#### Nombre de lignes et colonnes

df1.shape

---

#### Affichage du nom des colonnes

df1.columns

---

#### Suppression des colonnes non utlisées

df1 = df1.drop(["asins", "brand", "dimension", "ean", "keys", "manufacturer", "manufacturerNumber", 
                  "reviews.date", "reviews.dateSeen", "reviews.doRecommend", "reviews.numHelpful", "reviews.rating", 
                  "reviews.sourceURLs", "reviews.text", "reviews.title", "reviews.username", "upc", "weight"], axis=1)

df1.shape

---

#### Modification de la position des colonnes

df1 = df1[['id', 'name', 'colors', 'primaryCategories', 'categories', 'dateAdded', 'dateUpdated', 'sourceURLs', 'imageURLs']]

df1.columns

---

#### *Document numéro 2 : Datafiniti Electronics Products Pricing Data*

df2 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

#### Nombre de lignes et colonnes

df2.shape

---

#### Affichage du nom des colonnes

df2.columns

---

#### Suppression des colonnes non utlisées

df2 = df2.drop(["prices.amountMax", "prices.amountMin", "prices.availability", "prices.condition", 
                   "prices.currency", "prices.dateSeen", "prices.isSale", "prices.merchant", "prices.shipping", 
                   "prices.sourceURLs", "asins", "brand", "ean", "keys", "manufacturer", "manufacturerNumber", 
                   "upc", "weight", "Unnamed: 26", "Unnamed: 27", "Unnamed: 28", "Unnamed: 29", "Unnamed: 30"], axis=1)

df2.shape

---

#### Modification de la position des colonnes

df2 = df2[['id', 'name', 'primaryCategories', 'categories', 'dateAdded', 'dateUpdated', 'sourceURLs', 'imageURLs']]

df2.columns

---

#### Concatenation des deux dataframes

product = df1.append(df2, ignore_index=True, sort=False)

product.info()

product.isnull().sum()

---

#### Suppression des doublons

product = product.drop_duplicates('id')

product.info()

---

#### Remplacement des valeurs NULL dans "colors" par "Inconnue"

product["colors"].fillna("Inconnue", inplace = True)

product.info()

---

#### Modification de la position des colonnes, des noms, puis export en CSV

product = product[["id", "name", "colors", "primaryCategories", "categories", "dateAdded", "dateUpdated", "sourceURLs", "imageURLs"]]

product.rename(columns={"id": "product_id"}, inplace=True)

product.to_csv('./CSV_Tables/product.csv', sep=',', encoding='utf-8', index=False)

---

## 2 - Création du CSV pour la table "Manufacturer"

#### *Document numéro 1 : Datafinity Electronics Product Data*

df3 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

#### Affichage du nom des colonnes

df3.columns

---

#### Suppression des colonnes non utlisées

df3 = df3.drop(["asins", "categories", "colors", "dateAdded", "dateUpdated", 
                "dimension", "ean", "imageURLs", "keys", "name", "primaryCategories", 
                "reviews.date", "reviews.dateSeen", "reviews.doRecommend", "reviews.numHelpful", 
                "reviews.rating", "reviews.sourceURLs", "reviews.text", "reviews.title", 
                "reviews.username", "sourceURLs", "upc", "weight"], axis=1)

df3.columns

---

#### *Document numéro 2 : Datafiniti Electronics Products Pricing Data*

df4 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

#### Affichage du nom des colonnes

df4.columns

---

#### Suppression des colonnes non utlisées

df4 = df4.drop(["prices.amountMax", "prices.amountMin", "prices.availability", "prices.condition", 
                "prices.currency", "prices.dateSeen", "prices.isSale", "prices.merchant", "prices.shipping", 
                "prices.sourceURLs", "asins", "categories", "dateAdded", "dateUpdated", "ean", "imageURLs", "keys", 
                "name", "primaryCategories", "sourceURLs", "upc", "weight", "Unnamed: 26", "Unnamed: 27", "Unnamed: 28", 
                "Unnamed: 29", "Unnamed: 30"], axis=1)

df4.columns

---

#### Concatenation des deux dataframes

manufacturer = df3.append(df4, ignore_index=True, sort=False)

manufacturer.info()

---

#### Suppression des doublons

manufacturer = manufacturer.drop_duplicates()

manufacturer.info()

---

#### Remplacement des valeurs NULL dans "manufacturer" par "Inconnu"

manufacturer["manufacturer"].fillna("Inconnu", inplace = True)

manufacturer.info()

---

#### Modification de la position des colonnes, des noms, puis export en CSV

manufacturer = manufacturer[['manufacturerNumber', 'manufacturer', 'brand', 'id']]

manufacturer.rename(columns={"id": "product_id"}, inplace=True)

manufacturer.to_csv('./CSV_Tables/manufacturer.csv', sep=',', encoding='utf-8', index=False)

---

## 3 - Création du CSV pour la table "Logistics"

#### *Document numéro 1 : Datafinity Electronics Product Data*

df5 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

#### Affichage du nom des colonnes

df5.columns

---

#### Suppression des colonnes non utlisées

df5 = df5.drop(["brand", "categories", "colors", "dateAdded", "dateUpdated", "imageURLs", 
                "manufacturer", "manufacturerNumber", "name", "primaryCategories", "reviews.date", 
                "reviews.dateSeen", "reviews.doRecommend", "reviews.numHelpful", "reviews.rating", 
                "reviews.sourceURLs", "reviews.text", "reviews.title", "reviews.username", "sourceURLs"], axis=1)

df5.columns

---

#### *Document numéro 2 : Datafiniti Electronics Products Pricing Data*

df6 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

#### Affichage du nom des colonnes

df6.columns

---

#### Suppression des colonnes non utlisées

df6 = df6.drop(['prices.amountMax', 'prices.amountMin', 'prices.availability','prices.condition', 
                'prices.currency', 'prices.dateSeen','prices.isSale', 'prices.merchant', 'prices.shipping',
                'prices.sourceURLs', 'brand', 'categories', 'dateAdded','dateUpdated', 'imageURLs', 'manufacturer',
                'manufacturerNumber', 'name', 'primaryCategories', 'sourceURLs', 'Unnamed: 26', 'Unnamed: 27', 'Unnamed: 28', 
                'Unnamed: 29', 'Unnamed: 30'], axis=1)

df6.columns

---

#### Concatenation des deux dataframes

logistics = df5.append(df6, ignore_index=True, sort=False)

logistics.info()

---

#### Suppression des doublons

logistics = logistics.drop_duplicates()

logistics.info()

---

#### Remplacement des valeurs NULL dans "dimension" par "Inconnue"

logistics["dimension"].fillna("Inconnue", inplace = True)

logistics.info()

---

#### Modification de la position des colonnes, des noms, puis export en CSV

logistics = logistics[['asins', 'ean', 'upc', 'dimension', 'weight', 'keys', 'id']]

logistics.rename(columns={"id": "product_id"}, inplace=True)

logistics.to_csv('./CSV_Tables/logistics.csv', sep=',', encoding='utf-8', index=False)

---

## 4 - Création du CSV pour la table "Reviews"

#### *Document numéro 1 : Datafinity Electronics Product Data*

df7 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

#### Affichage du nom des colonnes

df7.columns

---

#### Suppression des colonnes non utlisées

df7 = df7.drop(["asins", "brand", "categories", "colors", "dateAdded", "dateUpdated", "dimension", "ean", 
                "imageURLs", "keys", "manufacturer","manufacturerNumber", "name", "primaryCategories", 
                "sourceURLs", "upc", "weight"], axis=1)

df7.columns

---

#### *Document numéro 2 : Datafiniti Electronics Products Pricing Data*

df8 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

#### Affichage du nom des colonnes

df8.columns

---

#### Suppression des colonnes non utlisées

df8 = df8.drop(["prices.amountMax", "prices.amountMin", "prices.availability", "prices.condition", "prices.currency", 
                "prices.dateSeen", "prices.isSale", "prices.merchant", "prices.shipping", "prices.sourceURLs", "asins", 
                "brand", "categories", "dateAdded", "dateUpdated", "ean", "imageURLs", "keys", "manufacturer", "manufacturerNumber", 
                "name", "primaryCategories", "sourceURLs", "upc", "weight", "Unnamed: 26", "Unnamed: 27", "Unnamed: 28", "Unnamed: 29",
                "Unnamed: 30"], axis=1)

df8.columns

---

#### Concatenation des deux dataframes

reviews = df7.append(df8, ignore_index=True, sort=False)

reviews.info()

---

#### Suppression des doublons

reviews = reviews.drop_duplicates()

reviews.info()

---

#### Modification de la position des colonnes, des noms, puis export en CSV

reviews = reviews[["id", "reviews.username", "reviews.date", "reviews.title", "reviews.text", "reviews.rating", 
                   "reviews.doRecommend", "reviews.numHelpful", "reviews.dateSeen", "reviews.sourceURLs"]]

reviews.rename(columns={"id": "product_id", "reviews.username": "username", "reviews.date": "date", "reviews.title": "title", 
                        "reviews.text": "text", "reviews.rating": "rating", "reviews.doRecommend": "doRecommend", 
                        "reviews.numHelpful": "numHelpful", "reviews.dateSeen": "dateSeen", "reviews.sourceURLs": "sourceURLs"},
               inplace=True)

Autre méthode : 
```sql 
reviews.columns.str.replace("reviews.", " ")
```


reviews.rating.unique()

reviews.to_csv('./CSV_Tables/reviews.csv', sep=',', encoding='utf-8', index=False)

---

## 5 - Création du CSV pour la table "Prices"

#### *Document numéro 1 : Datafinity Electronics Product Data*

df9 = pd.read_csv("./data/DatafinitiElectronicsProductData.csv")

---

#### Affichage du nom des colonnes

df9.columns

---

#### Suppression des colonnes non utlisées

df9 = df9.drop(["asins", "brand", "categories", "colors", "dateAdded", "dateUpdated", "dimension", 
                "ean", "imageURLs", "keys", "manufacturer", "manufacturerNumber", "name", "primaryCategories", 
                "reviews.date", "reviews.dateSeen", "reviews.doRecommend", "reviews.numHelpful", "reviews.rating", 
                "reviews.sourceURLs", "reviews.text", "reviews.title", "reviews.username", "sourceURLs", "upc", "weight"], axis=1)

df9.columns

---

#### *Document numéro 2 : Datafiniti Electronics Products Pricing Data*

df10 = pd.read_csv("./data/DatafinitiElectronicsProductsPricingData.csv")

---

#### Affichage du nom des colonnes

df10.columns

---

#### Suppression des colonnes non utlisées

df10 = df10.drop(["asins", "brand", "categories", "dateAdded","dateUpdated", "ean", 
                  "imageURLs", "keys", "manufacturer", "manufacturerNumber", "name", 
                  "primaryCategories", "sourceURLs", "upc", "weight", "Unnamed: 26", 
                  "Unnamed: 27", "Unnamed: 28", "Unnamed: 29", "Unnamed: 30"], axis=1)

df10.columns

---

#### Concatenation des deux dataframes

prices = df9.append(df10, ignore_index=True, sort=False)

prices.info()

---

#### Suppression des doublons

prices = prices.drop_duplicates()

prices.info()

---

#### Modification de la position des colonnes, des noms, puis export en CSV

prices = prices[["prices.amountMax", "prices.amountMin", "prices.availability", "prices.condition", "prices.currency", 
                 "prices.dateSeen", "prices.isSale", "prices.merchant", "prices.shipping", "prices.sourceURLs", "id"]]

prices.rename(columns={"prices.amountMax": "amountMax", "prices.amountMin": "amountMin", "prices.availability": "availability", 
                       "prices.condition": "condition", "prices.currency": "currency", "prices.dateSeen": "dateSeen", 
                       "prices.isSale": "isSale", "prices.Merchant": "Merchant", "prices.shipping": "shipping", 
                       "prices.sourceURLs": "sourceURLs", "id": "product_id"}, inplace=True)

prices.to_csv('./CSV_Tables/prices.csv', sep=',', encoding='utf-8', index=False)

---
