==> 04_Create_DB.ipynb <==
## (04) Création de la BDD, des tables et import des données sur PostgreSQL

---

## 01 - Création de la BDD

import psycopg2, os

conn = psycopg2.connect(database="postgres", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''CREATE database db_exac''';

cursor.execute(sql)
print("Base de données créée avec succès !")

conn.close()

---

## 02 - Création des tables

def creer_table(conn, sql_creation_table):
    try:
        cursor = conn.cursor()
        cursor.execute(sql_creation_table)
        conn.commit()
    except psycopg2.Error as e:
        print("Erreur lors de la création de la table")
        print(e)
        return
    cursor.close()
    print("Les tables ont été crées avec succès")

sql_creer_tables = """
    CREATE TABLE IF NOT EXISTS product (
    product_id CHAR(20) NOT NULL PRIMARY KEY,
    name VARCHAR(200),
    colors VARCHAR,
    primaryCategories VARCHAR,
    categories VARCHAR,
    dateAdded CHAR(20),
    dateUpdated CHAR(20),
    sourceURLs VARCHAR,
    imageURLs VARCHAR,
    review_id INT,
    price_id INT);
    
    CREATE TABLE IF NOT EXISTS manufacturer (
    manufacturer_id SERIAL PRIMARY KEY,
    manufacturer VARCHAR(40),
    brand VARCHAR,
    manufacturerNumber VARCHAR(30),
    product_id CHAR(20) NOT NULL,
    FOREIGN KEY (product_id) REFERENCES product(product_id));
    
    CREATE TABLE IF NOT EXISTS logistics (
    logistics_id SERIAL PRIMARY KEY,
    asins VARCHAR,
    ean VARCHAR,
    upc VARCHAR,
    dimension VARCHAR(50),
    weight VARCHAR,
    keys VARCHAR,
    product_id CHAR(20) NOT NULL,
    FOREIGN KEY (product_id) REFERENCES product(product_id));
    
    CREATE TABLE IF NOT EXISTS reviews (
    review_id SERIAL PRIMARY KEY,
    product_id CHAR(20) NOT NULL,
    username VARCHAR(40),
    date VARCHAR,
    title VARCHAR,
    text VARCHAR,
    rating REAL,
    doRecommend VARCHAR(20),
    numHelpful REAL,
    dateSeen VARCHAR,
    sourceURLs VARCHAR,
    FOREIGN KEY (product_id) REFERENCES product(product_id));
    
    CREATE TABLE IF NOT EXISTS prices (
    price_id SERIAL PRIMARY KEY,
    amountMax REAL,
    amountMin REAL,
    availability VARCHAR(20),
    condition VARCHAR,
    currency CHAR(3),
    dateSeen VARCHAR,
    isSale BOOL,
    merchant VARCHAR(50),
    shipping VARCHAR(50),
    sourceURLs VARCHAR,
    product_id CHAR(20) NOT NULL,
    FOREIGN KEY (product_id) REFERENCES product(product_id));
"""

def ouvrir_connection(nom_bdd, utilisateur, mot_passe, host='localhost', port=5432):
    
    try:
        conn = psycopg2.connect(dbname=nom_bdd, user=utilisateur, password=mot_passe, host=host, port=5432)
    except psycopg2.Error as e:
        print("Erreur lors de la connection à la base de données")
        print(e)
        return None
    conn.set_session(autocommit=True)
    
    return conn

ma_base_donnees = "db_exac"
utilisateur = "postgres"
mot_passe = os.environ.get("pg_psw")

conn = ouvrir_connection(ma_base_donnees, utilisateur, mot_passe)
type(conn)

creer_table(conn, sql_creer_tables)

---

## 03 - Insertion des données par le biais des fichiers CSV

### Importation données table "product"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''COPY product (product_id, name, colors, primaryCategories, categories, dateAdded, 
    dateUpdated, sourceURLs, imageURLs) 
    FROM 'C:\\Users\\Public\\CSV_Tables\\product.csv' WITH CSV HEADER 
    ENCODING 'utf8' DELIMITER ',' QUOTE '"';
    '''

cursor.execute(sql)
print("Importation réalisée avec succès !")

conn.close()

---

### Importation données table "manufacturer"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''COPY manufacturer (manufacturer, brand, manufacturerNumber, product_id) 
    FROM 'C:\\Users\\Public\\CSV_Tables\\manufacturer.csv' WITH CSV HEADER 
    ENCODING 'utf8' DELIMITER ',' QUOTE '"';
    '''

cursor.execute(sql)
print("Importation réalisée avec succès !")

conn.close()

---

### Importation données table "logistics"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''COPY logistics (asins, ean, upc, dimension, weight, keys, product_id) 
    FROM 'C:\\Users\\Public\\CSV_Tables\\logistics.csv' WITH CSV HEADER 
    ENCODING 'utf8' DELIMITER ',' QUOTE '"';
    '''

cursor.execute(sql)
print("Importation réalisée avec succès !")

conn.close()

---

### Importation données table "reviews"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''COPY reviews (product_id, username, date, title, text, rating, 
    doRecommend, numHelpful, dateSeen, sourceURLs) 
    FROM 'C:\\Users\\Public\\CSV_Tables\\reviews.csv' WITH CSV HEADER 
    ENCODING 'utf8' DELIMITER ',' QUOTE '"';
    '''

cursor.execute(sql)
print("Importation réalisée avec succès !")

conn.close()

---

### Importation données table "prices"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''COPY prices (amountMax, amountMin, availability, condition, 
    currency, dateSeen, isSale, merchant, shipping, sourceURLs, product_id) 
    FROM 'C:\\Users\\Public\\CSV_Tables\\prices.csv' WITH CSV HEADER 
    ENCODING 'utf8' DELIMITER ',' QUOTE '"';
    '''

cursor.execute(sql)
print("Importation réalisée avec succès !")

conn.close()

---

## 04 - Ajout des index créés pour "prices" et "reviews" dans la table "product"

### Ajout données "reviews"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''UPDATE product
         SET review_id = reviews.review_id
         FROM reviews
         WHERE reviews.product_id = product.product_id;
    '''

cursor.execute(sql)
print("Ajout effectué avec succès !")

conn.close()

---

### Ajout données "prices"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''UPDATE product
         SET price_id = prices.price_id
         FROM prices
         WHERE prices.product_id = product.product_id;
    '''

cursor.execute(sql)
print("Ajout effectué avec succès !")

conn.close()

---

## 05 - Ajout des FOREIGN KEY "price_id" et "review_id" dans la table "product"

### Ajout relation "reviews"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''ALTER TABLE product ADD FOREIGN KEY (review_id) REFERENCES reviews(review_id);
    '''

cursor.execute(sql)
print("Relation créée avec respect et robustesse !")

conn.close()

---

### Ajout relation "prices"

conn = psycopg2.connect(database="db_exac", user='postgres', password=os.environ.get("pg_psw"), host='localhost', port= '5432')

conn.autocommit = True

cursor = conn.cursor()

sql = '''ALTER TABLE product ADD FOREIGN KEY (price_id) REFERENCES prices(price_id);
    '''

cursor.execute(sql)
print("Relation créée avec respect et robustesse !")

conn.close()

---
